requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning - Core
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Interpretability
shap>=0.43.0

# Configuration and Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.65.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
